+++ tests/fixtures/code_objects/blocksworld_fond.py
@@ -13,12 +13,13 @@
 """This test module contains the fixtures for 'blocksworld-ipc08' domain and problem."""
 import pytest
 
-from pddl.core import Action, Domain, Problem, Requirements
+from pddl.core import Action, Domain, Problem
 from pddl.logic import Constant
 from pddl.logic.base import And, OneOf
 from pddl.logic.effects import AndEffect, When
 from pddl.logic.helpers import constants, variables
 from pddl.logic.predicates import EqualTo, Predicate
+from pddl.requirements import Requirements
 
 
 @pytest.fixture(scope="session")
