+++ tests/fixtures/code_objects/triangle_tireworld.py
@@ -13,11 +13,12 @@
 """This test module contains the fixtures for 'triangle-tireworld' domain and problem."""
 import pytest
 
-from pddl.core import Action, Domain, Problem, Requirements
+from pddl.core import Action, Domain, Problem
 from pddl.logic.base import And, OneOf
 from pddl.logic.effects import AndEffect
 from pddl.logic.helpers import constants, variables
 from pddl.logic.predicates import Predicate
+from pddl.requirements import Requirements
 
 
 @pytest.fixture(scope="session")
